backend/requirements.txt
Django==5.0.7
djangorestframework
django-cors-headers
djangorestframework-simplejwt